+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add chest to each dungeon with unique loot
 - Added a new option in the graphics menu to enable GPU timing (not always supported). The timing values can be viewed in the HUD debug info (F3) and will be saved as chrome trace files in the working directory when taking a screenshot.
 - Added new Present Mode option in the graphics menu. Selecting Fifo (i.e. vsync) or Mailbox can be used to eliminate screen tearing.
+- Quality color indicators next to recipe names in crafting menu
 
 ### Changed
 
@@ -121,6 +122,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved the rest of screenshot work into the background. Screenshoting no longer induces large pauses.
 - Reworked tidal warrior to have unique attacks
 - Reworked yeti to have unique attacks
+- Widened recipe name list in crafting menu
 
 ### Removed